requirements.txt
discord.py>=2.3.2
redis>=5.0.0
google-cloud-aiplatform
openai
python-dotenv